--HG--
rename : devtools/client/devtools-startup.js => devtools/shim/devtools-startup.js
rename : devtools/client/devtools-startup.manifest => devtools/shim/devtools-startup.manifest

+++ layout/style/ImageLoader.cpp
@@ -334,7 +334,7 @@ void InvalidateImagesCallback(nsIFrame* aFrame,
 
   if (nsLayoutUtils::InvalidationDebuggingIsEnabled()) {
     printf_stderr("Invalidating display item(type=%d) based on frame %p \
-      because it might contain an invalidated image\n", type, aFrame);
+      because it might contain an invalidated image\n", static_cast<uint32_t>(type), aFrame);
   }
   aItem->Invalidate();
   aFrame->SchedulePaint();
